src/datalayer/models/base.py
from tortoise import fields
from tortoise.models import Model

class ModelBase(Model):
    id = fields.IntField(primary_key=True)
    created_at = fields.DatetimeField(auto_now_add=True)
    modified_at = fields.DatetimeField(auto_now=True)
