pyproject.toml
[project]
name = "rede-social-com-python"
version = "0.1.0"
description = ""
authors = [
    {name = "Eli5eu",email = "eli5euc4mb1nd4@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "tortoise-orm[asyncpg] (>=0.25.1,<0.26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.14.1"

[tool.taskipy.tasks]
dev = "fastapi dev dev.py"
prod = "uvicorn src.api.application:app --host 0.0.0.0 --port 8000 --reload"
